refactor: condense demo directories 👍
