--- demo/components/ComponentsScreen/ActiveScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Active.md'
