--- demo/components/ComponentsScreen/AlertsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Alerts.md'
