--- demo/components/ComponentsScreen/ButtonsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Buttons.md'
