--- demo/components/ComponentsScreen/CardsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Cards.md'
