+++ demo/components/ComponentsScreen/ComponentsScreen.jsx
@@ -7,18 +7,18 @@ import GroupNav from 'components/universal/GroupNav'
 import componentRoutes from 'utils/component-routes'
 import Header from 'components/universal/Header'
 
-import ActiveScreen from './ActiveScreen'
-import AlertsScreen from './AlertsScreen'
-import ButtonsScreen from './ButtonsScreen'
-import CardsScreen from './CardsScreen'
-import DrawersScreen from './DrawersScreen'
-import DropdownsScreen from './DropdownsScreen'
-import ListGroupsScreen from './ListGroupsScreen'
-import ModalsScreen from './ModalsScreen'
-import NavsScreen from './NavsScreen'
-import PopoversScreen from './PopoversScreen'
-import TabsScreen from './TabsScreen'
-import TooltipsScreen from './TooltipsScreen'
+import ActiveScreen from './Screens/Active.md'
+import AlertsScreen from './Screens/Alerts.md'
+import ButtonsScreen from './Screens/Buttons.md'
+import CardsScreen from './Screens/Cards.md'
+import DrawersScreen from './Screens/Drawers.md'
+import DropdownsScreen from './Screens/Dropdowns.md'
+import ListGroupsScreen from './Screens/ListGroups.md'
+import ModalsScreen from './Screens/Modals.md'
+import NavsScreen from './Screens/Navs.md'
+import PopoversScreen from './Screens/Popovers.md'
+import TabsScreen from './Screens/Tabs.md'
+import TooltipsScreen from './Screens/Tooltips.md'
 
 import { component } from './components-screen.scss'
