--- demo/components/ComponentsScreen/DrawersScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Drawers.md'
