--- demo/components/ComponentsScreen/DropdownsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Dropdowns.md'
