--- demo/components/ComponentsScreen/ListGroupsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './ListGroups.md'
