--- demo/components/ComponentsScreen/ModalsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Modals.md'
