--- demo/components/ComponentsScreen/NavsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Navs.md'
