--- demo/components/ComponentsScreen/PopoversScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Popovers.md'
