--- demo/components/ComponentsScreen/TabsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Tabs.md'
