--- demo/components/ComponentsScreen/TooltipsScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Tooltips.md'
