+++ demo/components/ConceptsScreen/ConceptsScreen.jsx
@@ -10,7 +10,7 @@ import ArchitectureScreen from 'GUIDES/Architecture.md'
 import AccessibilityScreen from 'GUIDES/Accessibility.md'
 import ComponentsScreen from 'GUIDES/Components.md'
 import ThemingScreen from 'GUIDES/Theming.md'
-import OverviewScreen from './OverviewScreen'
+import OverviewScreen from './Screens/Overview.md'
 
 import { component } from './concepts-screen.scss'
