--- demo/components/ConceptsScreen/OverviewScreen/index.js
@@ -1,1 +1,0 @@
-export { default } from './Overview.md'
